codex - add reusable nav drawer
